% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asian.R
\name{geomAsian}
\alias{geomAsian}
\title{Returns vector of continuous geometric European Asian call and put
prices}
\usage{
geomAsian(s, k, v, r, tt, d, numavg)
}
\description{
Returns vector of continuous geometric European Asian call and put
prices
}

