% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asian.R
\name{geomAsianPut}
\alias{geomAsianPut}
\title{price of continuous geometric average European Asian put}
\usage{
geomAsianPut(s, k, v, r, tt, d, numavg)
}
\description{
price of continuous geometric average European Asian put
}

