% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetual.R
\name{perpetual}
\alias{callperpetual}
\alias{perpetual}
\alias{putperpetual}
\title{Perpetual American options}
\usage{
callperpetual(s, k, v, r, d, priceonly)
putperpetual(s, k, v, r, d, priceonly)
}
\arguments{
\item{s}{Stock price}

\item{k}{Strike price of the option}

\item{v}{Volatility of the stock, defined as the annualized
standard deviation of the continuously-compounded return}

\item{r}{Annual continuously-compounded risk-free interest rate}

\item{d}{Dividend yield, annualized, continuously-compounded}

\item{priceonly}{Boolean (TRUE). If false, the option price and
exercise barrier are returned as a list}
}
\value{
Option price, and optionally the optimal exercise barrier.
}
\description{
\code{callperpetual} and \code{putperpetual} compute
    prices of perpetual American options. The functions optionally
    return the exercise barriers (the prices at which the options
    are optimally exercised).

If the dividend yield is zero, a perpetual call is never exercised
and the value is zero. Similarly, if the risk-free rate is zero, a
perpetual put is never exercised and the value is zero
}
\details{
Returns a scalar or vector of option prices, depending on
the inputs

\code{callperpetual(s, k, v, r, tt, d)}
}
\examples{
s=40; k=40; v=0.30; r=0.08;  d=0.02;
callperpetual(s, k, v, r, d)

putperpetual(s, c(35, 40, 45), v, r, d, priceonly=FALSE)


}

